spectrum_monitor/requirements.txt
numpy>=1.19.0
PyYAML>=5.3.0
termcolor>=1.1.0
keyboard>=0.13.0 